Handling of absolute and relative times distinguished by flag.
Meanwhile added a proxy interface, so the traditional interface of
operating with two operands of rlTimeEx is available. 
One small exception is there of returning double for the difference 
of two rlTimeEx object by default and returning rlTimeEx objects only 
on request by casting the second operand to RelativeTime(rltime_object).