Reduced data loss on frequent logger restarts.
By inspecting the logging files on logger start the data loss from 
deleted ring buffer files is reduced. 

This is achieved by avoiding the deletion of the oldest of the ring 
buffer files in case the youngest file is still in time window for 
further logging events.

This way occurred a problem in Windows, where read and write pointer are 
not independent. So the inspection of the files and writing to the files 
is done from different file descriptors, opend for the special purpose. 

By the way reduced this problems with data logging, reopening the log
files etc. after a server crash and restart.